Desktop/electron game crash

See merge request polytechnique-montr-al/log3900/21-1/equipe-207/log3900-207!267

Former-commit-id: be2c5072d03ff63e498a914941f3b4883be4cea0

+++ server/app/services/sockets/socket-connection.service.ts
@@ -17,7 +17,8 @@ import { UserService } from '../user.service';
 
 @injectable()
 export class SocketConnectionService {
-
+    private socketIdLinks: Map<string, any> = new Map<string, any>();
+    private externalWindowSocketIds: string[] = [];
 
     constructor(
         @inject(TYPES.SocketService) private socketService: SocketService,
@@ -46,6 +47,20 @@ export class SocketConnectionService {
                 } catch (err) {
                     console.log(err);
                 }
+            });
+
+            socket.on('openExternalChat', (data: any) => {
+                if(data.isExternalWindow){
+                    this.externalWindowSocketIds.push(socket.id);
+                }
+                this.socketIdLinks.set(data.linkedSocketId, socket);
+                this.socketService.getSocket().to(data.linkedSocketId).emit('openExternalChatCallback', { "externalWindowSocketid": socket.id });
+            });
+
+            socket.on('closeExternalChat', (data: any) => {
+                let externalSocket = this.socketIdLinks.get(data.mainWindowId);
+                this.socketIdLinks.delete(data.mainWindowId);
+                if(externalSocket) this.socketIdLinks.delete(externalSocket.id);
             });
 
             socket.on('drawingEvent', (drawingEvent: DrawingEvent) => {
@@ -166,6 +181,11 @@ export class SocketConnectionService {
                     await this.userService.addUserToChat(user.username, request.chatId)
                     await this.chatManagerService.addUserToChat(user.username, request.chatId)
                     this.socketService.getSocket().to(socket.id).emit('joinChatRoomCallback');
+                    if(this.socketIdLinks.has(socket.id)){
+                        let linkedSocket = this.socketIdLinks.get(socket.id);
+                        linkedSocket.join(request.chatId);
+                        this.socketService.getSocket().to(linkedSocket.id).emit('refreshChatRequest', {chatId: request.chatId});
+                    }
                 }
                 catch (e) {
                     console.error(e);
@@ -182,6 +202,11 @@ export class SocketConnectionService {
                     await this.userService.removeUserFromChat(user.username, request.chatId)
                     await this.chatManagerService.removeUserFromChat(user.username, request.chatId)
                     this.socketService.getSocket().to(socket.id).emit('leaveChatRoomCallback');
+                    if(this.socketIdLinks.has(socket.id)){
+                        let linkedSocket = this.socketIdLinks.get(socket.id);
+                        linkedSocket.leave(request.chatId);
+                        this.socketService.getSocket().to(linkedSocket.id).emit('refreshChatRequest', {chatId: "General"});
+                    }
                 }
                 catch (e) {
                     console.error(e);
@@ -190,16 +215,25 @@ export class SocketConnectionService {
 
             socket.on('disconnect', () => {
                 try {
-                    const user: any = this.tokenService.getTokenInfo(socket.handshake.query.authorization);
-                    this.authService.addUserToLogCollection(user.username, false);
-                    console.log('disconnection of ' + user.username);
-                    const lobbyId = this.lobbyManagerService.isUserInLobby(user.username);
-                    if (lobbyId) {
-                        this.lobbyManagerService.removePlayerFromLobby(user, lobbyId)
+                    if(!this.externalWindowSocketIds.includes(socket.id)){
+                        const user: any = this.tokenService.getTokenInfo(socket.handshake.query.authorization);
+                        this.authService.addUserToLogCollection(user.username, false);
+                        console.log('disconnection of ' + user.username);
+                        const lobbyId = this.lobbyManagerService.isUserInLobby(user.username);
+                        if (lobbyId) {
+                            this.lobbyManagerService.removePlayerFromLobby(user, lobbyId)
+                        }
+                        const gameId = this.gameManagerService.isUserInGame(user.username);
+                        if (gameId) {
+                            this.gameManagerService.disconnectGame(gameId, user.username);
+                        }
                     }
-                    const gameId = this.gameManagerService.isUserInGame(user.username);
-                    if (gameId) {
-                        this.gameManagerService.disconnectGame(gameId, user.username);
+                    else{
+                        for( let i = 0; i < this.externalWindowSocketIds.length; i++){ 
+                            if ( this.externalWindowSocketIds[i] === socket.id) { 
+                                this.externalWindowSocketIds.splice(i, 1); 
+                            }
+                        }
                     }
                 }
                 catch (e) {
